Dynamic entity removal
Prior to this point, drawable meshes could only be added to the scene, as there was no way for the graphics frontend to communicate to the backend that "hey, this entity ID is unused now, let's use it in the future".

This commit changes the way the frontend & backend communicate with the shared transform buffer, which allows for dynamic entity removal and saves a lot of memory in the process!