Changed Tagging System
The Grove's system for managing game entities allows for "tagging" special entities with a unique identifier, so that they can later be accessed from other code using that same identifier.

Previously, identifiers were plain strings (i.e. 'player', 'camera').  This approach is error-prone, as it is possible to misspell the identifier when looking up or setting a tag.  Also, looking up entities via a tag creates a tight dependency to the code that creates that entity.  If tags are strings, it can be very difficult to find these dependencies later down the line, because it doesn't require an explicit `import` statement between files.

The newer system uses JavaScript's `Symbol` type for tagging entities.  This ensures that typos will be caught at compile-time, not run-time, as well as creating explicit dependencies between modules, as the symbol must be imported before use.