Shooting was GameScript, becomes utility method
`src/js/game/shooting.ts` was predominantly used for its `shoot` method, but it also exported a GameScript to add click-to-shoot functionality.  Given the dependencies of that functionality, that script was better placed in the player GameScript.  Now, `shooting.ts` only exports the `shoot` method, which creates a ball and launches it in a given direction.