Control game loop through event bus
Manually reaching into the guts of the engine to switch a "running" boolean seemed like a really dirty way of pausing and playing the game.  With this new implementation, you can send (and listen to!) start/stop events.  I acknowledge that these are very vaguely named, and "start" really means "play" and "stop" really means "pause".  It's not an entire architectural change per se, it just seems like a better way of starting/stopping the core loop than switching the master boolean manually.