Realized asset loading was fucked, and un-fucked it (also swords)
In my journey to port the old Grove game into the new engine, I wanted to start with melee combat as it seemed the most accessible feature to port.  I made a game script, loaded the sword model, and had it rotate down in a swing motion on every mouseclick event.

Lo and behold, only the last mesh in the model hierarchy would rotate.  This was because of some funk in the design of the asset loader, and how it dispatched callbacks when assets were loaded.

By the end of this commit, I had accidentally introduced a bug where the game map would be added to the scene 9,0000 times, but caught it and cut that back to one (WOO for memory consumption!)

Also, there are no longer separate functions for callback-based and promise-based asset loading: everything is promise-based, and it works REGARDLESS of how many children meshes are in the model.

Overall, this was a really good commit, and a good way to end the night :smile: