Added UI, refactored graphics frontend/backend
UI elements such as the crosshair and the player's HUD are now rendered via a separate ThreeJS scene and camera, which means A) less dependency on the DOM (crosshair) and B) UI elements no longer clip through meshes (hud) because they use a separate scene.

Doubled down on the command-based graphics architecture, as is visible from `commands.ts`.

In fixing the asset loading, I seemed to have broken object deletion on the graphics side.  TODO: fix this.