Fixed pesky deletion bug
Found that nasty lil' graphics bug I talked about in the last commit.  The issue was several-fold.

To start, since the frontend recursively submits all meshes in a model to the backend when that model is added to the scene, then when that model is removed from the scene it also needs to have all of its children removed recursively.  Prior, I was submitting all children meshes to the backend individually but removing the whole model all at once on deletion - which is a no-go.

After making that change, I found that deletion was still broken, but it was giving me a handy error.  After **several** (and I mean many) `console.log` debugging trials, I found that the backend was receiving an `undefined` mesh ID when it came time to delete a model.  Why was this happening?

Simple: meshes were being deleted recursively, but the `id` being submitted was still the `id` of the _entire model_.  This was typically a `Group`, or something non-meshy, so there was no actual ID associated with it and the backend wasn't getting the correct data.  Changing `object.userData.meshId` to `node.userData.meshId` in `graphics.ts` solved the issue.