Fixed graphics memory consumption
It turns out that ThreeJS (de)serialization is capable of reproducing parent/child relationships, so the recursive technique I used in prior commits was actually unneccessary and wasteful: I would submit a parent node to the backend (which would include all of its children), as well as submitting each of its children individually.  The technique currently used is to only submit the topmost parent node when a mesh or model is added, and just recursively send all texture information to the backend.  From there, the backend would recursively destructure all material/texture information.  So, _hopefully_, each renderable only exists once on the backend now.