Begin migration to Ammo.js
My eventual goal is to move all physics simulation to a C++ addon running in the electron main process, and use shared memory to give the JavaScript code access to object transforms.  That way, The Grove will get all the physics performance of a native game, and the main game scripts will still be written in JS/TS.

As part of that transition, since the C++ code will use Bullet3, I'm gradually switching to using Ammo.js instead of Cannon.js, since Ammo is a direct Emscripten port of bullet3.